requirements-dev.txt
# Testing - includes pytest, pytest-asyncio, pytest-cov, and Home Assistant dependencies
pytest-homeassistant-custom-component==0.13.163

# Integration-specific dependencies
av==16.0.1

# Linting and Formatting
ruff==0.8.4

# Type checking
mypy==1.13.0
